+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed unsealed TAPSIGNER being detected as already initialized
 - Show full address derivation path including the keychain index
 - Fix crashing on APDU NFC error
+- Fix verifying words flow not working if seed has a consecutive duplicate word
 
 ## [1.0.0] - 2025-06-11